requirements.txt
python-telegram-bot==20.7
feedparser==6.0.11
requests==2.31.0
